backend/requirements.txt
# Workspace version: 1.46.40
ipykernel==6.29.5
python-multipart==0.0.9
uvicorn[standard]==0.34.2
databutton==0.38.34
fastapi==0.115.12
openai
beautifulsoup4
requests
firebase-admin
mysql-connector-python
databutton==0.38.34
ipykernel==6.29.5
fastapi==0.111.0
python-multipart==0.0.9
uvicorn[standard]==0.29.0
pandas
openpyxl
apscheduler